refactor(test): create setup package for base test classes
- manage AcceptanceTest and DatabaseTest setup formats
