+++ src/test/java/woowacourse/shoppingcart/application/CustomerServiceTest.java
@@ -12,7 +12,7 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import woowacourse.shoppingcart.domain.Customer;
 import woowacourse.auth.domain.User;
 import woowacourse.shoppingcart.dao.CustomerDao;
-import woowacourse.shoppingcart.dao.DatabaseTest;
+import woowacourse.setup.DatabaseTest;
 import woowacourse.shoppingcart.dto.request.SignUpRequest;
 import woowacourse.shoppingcart.dto.request.UpdateMeRequest;
 import woowacourse.shoppingcart.dto.request.UpdatePasswordRequest;
